Home page image sizing
Having `.jumbotron-container` with `display: flex;` will automatically make its direct child element(s) have the same height and width (if it's the only element) of the container. To proportionally scale, I converted the `height` and `width` of the image to `max-height` and `max-width`. Since the measurement of the container is based on viewport sizing, I also converted the measurements from percentages to viewport sizing.

This should scale well no matter the screen size, now!